+++ test/validation_sqlite_test.rb
@@ -4,73 +4,31 @@ require "simplecov"
 SimpleCov.start
 require "minitest/autorun"
 require_relative "../lib/hashy_validator"
-
-# Create database
-conn = { adapter: "sqlite3", database: ":memory:" }
-ActiveRecord::Base.establish_connection(conn)
-
-# Schema
-ActiveRecord::Schema.define do
-  create_table :profiles do |t|
-    t.string(:name)
-    t.string(:notifications)
-
-    # @Todo: use Postgres
-    # Also check if i really need to create test with pg integration
-    # t.jsonb :notifications
-  end
-  create_table :products do |t|
-    t.string(:name)
-    t.string(:discount_by_quantity)
-  end
-
-  create_table :items do |t|
-    t.string(:name)
-    t.string(:metas)
-  end
-end
-
-# Model
-class Profile < ActiveRecord::Base
-  validates :name, presence: true
-  validates :notifications, hashy_array: {
-    type: HashValidator.multiple("string", "unique"),
-    header: HashValidator::Validations::Optional.new("string"),
-  }
-end
-
-class Product < ActiveRecord::Base
-  validates :name, presence: true
-  validates :discount_by_quantity, hashy_array: {
-    uuid: "unique",
-    quantity: HashValidator.multiple("numeric"),
-    price: HashValidator.multiple("numeric"),
-    active: HashValidator.multiple("boolean"),
-  }
-end
-
-class Item < ActiveRecord::Base
-  validates :name, presence: true
-  validates :metas,
-    hashy_array: {
-      active: HashValidator.multiple("boolean"),
-    },
-    if: ->(item) { self.quantity > 1 }
-
-  attribute :quantity, :integer
-end
+require_relative "database/sqlite"
+require_relative "model/samples"
+require "pry"
 
 # I am using JSON.generate is because sqlite3 does not have the jsonb column type
 class HashyArrayValidationTest < Minitest::Test
+  # =================================
+  # Success cases
+  # =================================
+
   def test_valid_hashy_returns_success
     profile = Profile.new(name: "John Doe", notifications: JSON.generate(
       [
         {
           type: "something",
-          header: "something header"
+          header: "something header",
         },
       ],
     ))
+
+    assert(profile.valid?)
+  end
+
+  def test_ignore_empty_array_returns_success
+    profile = Profile.new(name: "John Doe", notifications: JSON.generate([]))
 
     assert(profile.valid?)
   end
@@ -95,6 +53,65 @@ class HashyArrayValidationTest < Minitest::Test
 
     assert(item.valid?)
   end
+
+  def test_valid_hashy_object_returns_success
+    customer = Customer.new(age: 23, custom: JSON.generate({
+      name: "John Doe",
+      quantity: 20,
+      active: true,
+    }))
+
+    assert(customer.valid?)
+  end
+
+  def test_missing_optional_attribute_on_hashy_object_returns_success
+    customer = Customer.new(age: 23, custom: JSON.generate({
+      quantity: 20,
+      active: true,
+    }))
+
+    assert(customer.valid?)
+  end
+
+  def test_empty_hashy_object_returns_success
+    customer = Customer.new(age: 23, custom: JSON.generate({}))
+
+    assert(customer.valid?)
+  end
+
+  def test_ignore_nil_hashy_array_returns_success
+    profile = Profile.new(name: "John Doe", notifications: nil)
+
+    assert(profile.valid?)
+  end
+
+  def test_ignore_nil_hashy_object_returns_success
+    customer = Customer.new(age: 23, custom: nil)
+
+    assert(customer.valid?)
+  end
+
+  def test_valid_hashy_object_passing_boolean_returns_success
+    customer = Customer.new(age: 23, custom: JSON.generate({
+      quantity: 1,
+      active: true,
+    }))
+
+    assert(customer.valid?)
+  end
+
+  def test_valid_hashy_object_passing_boolean_as_string_returns_success
+    customer = Customer.new(age: 23, custom: JSON.generate({
+      quantity: 1,
+      active: "true",
+    }))
+
+    assert(customer.valid?)
+  end
+
+  # =================================
+  # Failures cases
+  # =================================
 
   def test_valid_hashy_with_condition_returns_fail
     item = Item.new(name: "Go Pro", quantity: 2, metas: JSON.generate([{}]))
@@ -172,5 +189,36 @@ class HashyArrayValidationTest < Minitest::Test
     ])
 
     refute(profile.valid?)
+  end
+
+  def test_invalid_hashy_object_returns_error
+    customer = Customer.new(age: 23, custom: JSON.generate({
+      name: "John Doe",
+    }))
+
+    refute(customer.valid?)
+  end
+
+  def test_invalid_hashy_object_passing_array_returns_error
+    customer = Customer.new(age: 23, custom: JSON.generate([{
+      name: "John Doe",
+    }]))
+
+    refute(customer.valid?)
+  end
+
+  def test_invalid_hashy_object_passing_not_valid_parse_returns_error
+    customer = Customer.new(age: 23, custom: "not a hash")
+
+    refute(customer.valid?)
+  end
+
+  def test_hashy_object_passing_invalid_boolean_returns_success
+    customer = Customer.new(age: 23, custom: JSON.generate({
+      quantity: 1,
+      active: "invalid",
+    }))
+
+    refute(customer.valid?)
   end
 end